Rename : `comment` 패키지 시그니처 변경
