+++ backend/src/main/java/com/team6/issue_tracker/comment/CommentContents.java
@@ -1,11 +1,15 @@
-package com.team6.issue_tracker.comment.domain;
+package com.team6.issue_tracker.comment;
 
 import lombok.Getter;
 import lombok.Value;
 
+import javax.validation.constraints.NotNull;
+
 @Value
 @Getter
 public class CommentContents {
+
+    @NotNull
     String contents;
 
     public CommentContents(String contents) {
